+++ __init__.py
@@ -1,5 +1,0 @@
-"""Init file for blinkpy."""
-from blinkpy.blinkpy import Blink
-from blinkpy.blinkpy import BlinkAuthenticationException
-from blinkpy.blinkpy import BlinkException
-from blinkpy.blinkpy import BlinkURLHandler